Update and delete for Harvest

+++ src/main/java/me/elhakimi/citronix/rest/controller/HarvestController.java
@@ -4,15 +4,13 @@ package me.elhakimi.citronix.rest.controller;
 
 import lombok.AllArgsConstructor;
 import me.elhakimi.citronix.domain.Harvest;
+import me.elhakimi.citronix.domain.dto.HarvestDTO;
 import me.elhakimi.citronix.rest.vm.HarvestVm;
 import me.elhakimi.citronix.rest.vm.mapper.HarvestVmMapper;
 import me.elhakimi.citronix.service.impl.HarvestServiceImpl;
 import me.elhakimi.citronix.util.ResponseUtil;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -29,22 +27,32 @@ public class HarvestController {
     }
 
     @PostMapping
-    public ResponseEntity<Object> save(@RequestBody Harvest harvest) {
-            HarvestVm harvestVm = harvestVmMapper.toHarvestVm(harvestService.save(harvest));
-            return ResponseUtil.saveSuccessfully("Harvest" , harvestVm);
+    public ResponseEntity<Object> save(@RequestBody HarvestDTO harvestDto) {
+        HarvestDTO savedHarvest = harvestService.save(harvestDto);
+        HarvestVm harvestVm = harvestVmMapper.toHarvestVm(savedHarvest);
+        return ResponseUtil.saveSuccessfully("Harvest", harvestVm);
     }
 
     public Harvest getHarvest(Long id) {
         return harvestService.getHarvest(id);
     }
 
-    public void delete(Long id) {
-        harvestService.delete(id);
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Object> delete(@PathVariable Long id) {
+        if(harvestService.getHarvest(id) != null) {
+            harvestService.delete(id);
+            return ResponseUtil.deleteSuccessfully("Harvest");
+        }
+
+        return ResponseUtil.notFound("Harvest");
+
     }
 
-    public Harvest update(Harvest harvest) {
-        return harvestService.update(harvest);
+    @PutMapping
+    public ResponseEntity<Object> update(@RequestBody  HarvestDTO harvestDto) {
+        HarvestDTO updatedHarvest = harvestService.update(harvestDto);
+        HarvestVm harvestVm = harvestVmMapper.toHarvestVm(updatedHarvest);
+        return ResponseUtil.updateSuccessfully("Harvest", harvestVm);
     }
-
-
 }
